Abstract - interface example

+++ JDay_3/src/main/java/ex3/Shape.java
@@ -2,5 +2,5 @@ package ex3;
 
 public interface Shape {
 
-	public int getArea();
+	public double getArea();
 }
